requirements.txt
langchain_community
langchain_core
PyPDF2
python-dotenv
streamlit
pandas
faiss-CPU
